requirements.txt
# Core LangChain Dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.7
langchain-openai>=0.0.5
openai>=1.6.1

# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# HTTP Client for MCP
httpx>=0.25.0

# Data Validation
pydantic>=2.5.3

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Utilities
tiktoken>=0.5.2
